Split solution
